test: move unit tests to lib
